requirements.txt
mss==6.1.0
rx==1.6.1
PyQt5==5.15.9
pywinauto==0.6.8
Pillow==11.0.0
dxcam==0.0.5
Flask==2.0.3
requests==2.28.1
pynput==1.7.6
qt-material==2.14
pywin32==308
pytest==8.3.3
